requirements.txt
numpy
matplotlib
tensorflow
keras
torchvision
cx_Freeze
pyinstaller
